Devtools demonstration
This commit demonstrates some of the features of devtools. We can change
frontend and backend code and the app will update automatically. If we
are logged in, our session will be restored if the app has to be
restarted.

New in Spring Boot 1.4, adding an template in `/error` with a valid HTTP
status code will bind it automatically as an error page. This commit
creates a custom error page if a page is not found (404).