Externalize configuration
This commit moves the hard-coded list of admin users to a configurable
property using a type-safe `@ConfigurationProperties` POJO. This commit
also adds the annotation processor so that the relevant meta-data is
generated at build time for IDE assistance.

If we remove one of the expected admin from the list, you get a 403 so
we could create a dedicated error page for that. Since we want a general
client-side error page, we rename our 404 to `4xx` so that Boot uses it
for any client-side error by default.