Auto-configuration demonstration
This commit demonstrates how one can easily write an auto-configuration.

A common mistake is to put the auto-configuration in a package that is a
target for component scan. We make sure to place it outside of
"com.example.cfp" (auto-configurations are meant to be placed in their
own module and not in an application anyway).

The auto-configuration provides a GitHubClient so that the complexity
of dealing with the `RestTemplate` setup is taken care externally.

We also demonstrate how auto-configurations can be ordered. We need to
check that the `CounterService` bean is available but we can only do that
if its auto-configuration was applied.